Add acceptance test around the bahaviour of the :sign_up feature flip.
Tagged as @flip and @sign_up so that we can easily run all flips or just this
flip.

Written as a long running test that first turns the flip on, checks the
behaviour, then turn the flip off and checks the behaviour. I've written it
this way because the flip defaults to off and will be left in that state after
this feature runs.