Feature flip sign up, remove beta user list.
Instead of maintaining a list of beta users that are allowed to sign up, we
now just turn the whole feature off. We can turn it on for ourselves, and then
turn it back off to lock others out.

This results in a better tested feature, that is usable going forward without
make any changes to it.