SwitchBoard: A wrapper for the rollout gem.
SwitchBoard is a class that provides the ability to toggle features on and
off. It wraps the rollout gem so that we do not expose the gems API to our
code base.

This will make it easier to upgrade or change feature flipping gems in the
future should we desire to do so.

It also serves as a single place to document all known feature flips in use
within the application.