Switchboard now groks groups and users.
When activating or deactivating a flip, you can optionally pass in a group. This
will trigger rollouts activate_group method to be called. If no group is passed
in then we trigger roullouts activate method and the feature is available for
all users.

The :all group is not used as it requires a member to be present. i.e. the
sign_in flip doesn't require a member, just a user of the site.

The same applies for deactivate.

When checking if a feature is active, we now pass in the user so that we can
check if they are part of a given group or not.