Ensure we don't render headings without content.
This would occur if the feature flip for following was turned off. The View
Controllers would have activities under dates, and render headings, but the
follow and unfollow partials would render nothing.

Now we strip them out when requesting them, and the partials always render if
called. No more conditionals in partials. Yay!