Some meta programming and ohter niceties to allow us to interact with FIT files.
For each of the global message types, there is an associated *_records method
generated that will return an array of records. If for some reason the global
message has been defined twice, they will raise an exception. I don't expect it
to happen but I'd rather explode than find out the hard way when the data we
try and read out of local messages fails in strange ways.

The global message types are:

  * file_id
  * file_creator
  * software
  * file_capabilities
  * mesg_capabilities
  * field_capabilities
  * device_settings
  * user_profile
  * hrm_profile
  * sdm_profile
  * bike_profile
  * zones_target
  * sport
  * hr_zone
  * power_zone
  * met_zone
  * goal
  * activity
  * session
  * lap
  * record
  * event
  * device_info
  * course
  * course_point
  * workout
  * workout_step
  * totals
  * weight_scale
  * blood_pressure